@@ -1,7 +1,7 @@
 {
   "name": "pdir/contao-theme-helper-bundle",
   "type":"contao-bundle",
-  "description": "pdir theme helper bundle for Contao 4",
+  "description": "pdir theme helper bundle for Contao 5",
   "keywords":["contao","theme","template", "tools", "helper"],
   "homepage": "https://pdir.de",
   "license": "LGPL-3.0+",
@@ -19,8 +19,8 @@
     "source" : "https://github.com/pdir/contao-theme-helper-bundle"
   },
   "require": {
-    "php": "^7.0 || ^8.0",
-    "contao/core-bundle": "^4.4 || ^5.0"
+    "php": "^8.0",
+    "contao/core-bundle": "^5.0"
   },
   "conflict": {
     "contao/core": "*"
