@@ -18,7 +18,7 @@ namespace Pdir\ThemeHelperBundle\EventListener;
 
 use Contao\ArticleModel;
 use Contao\ContentElement;
-use Contao\CoreBundle\Framework\ContaoFrameworkInterface;
+use Contao\CoreBundle\Framework\ContaoFramework;
 use Contao\Events;
 use Contao\StringUtil;
 
@@ -30,7 +30,7 @@ use Contao\StringUtil;
 class InsertTagsListener
 {
     /**
-     * @var ContaoFrameworkInterface
+     * @var ContaoFramework
      */
     private $framework;
 
@@ -44,9 +44,9 @@ class InsertTagsListener
     /**
      * Constructor.
      *
-     * @param ContaoFrameworkInterface $framework
+     * @param ContaoFramework $framework
      */
-    public function __construct(ContaoFrameworkInterface $framework)
+    public function __construct(ContaoFramework $framework)
     {
         $this->framework = $framework;
     }
