@@ -1,5 +1,8 @@
 <?php
 
+use Contao\Backend;
+use Contao\PageModel;
+
 /**
  * Add fields to tl_content
  */
@@ -18,7 +21,7 @@ $GLOBALS['TL_DCA']['tl_article']['fields']['pdir_th_tag'] = array
     'label'         => &$GLOBALS['TL_LANG']['tl_article']['pdir_th_tag'],
     'inputType'     => 'select',
     'search'        => true,
-    'options'       => $GLOBALS['tl_config']['theme_tags'],
+    'options'       => $GLOBALS['tl_config']['theme_tags']?? [],
     'reference'     => &$GLOBALS['TL_LANG']['tl_article']['th_tags'],
     'eval'          => array('mandatory'=>false, 'maxlength'=>64, 'tl_class'=>'w50 wizard'),
     'sql' => "varchar(64) NOT NULL default ''"
